+++ app/src/main/java/uz/behzoddev/restaurantpos/data/local/datasource/ItemDataSource.kt
@@ -6,7 +6,7 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodItem
 interface ItemDataSource {
     suspend fun insertFoodItem(foodItem: FoodItem): Long
     suspend fun updateFoodItem(foodItem: FoodItem): Long
-    suspend fun deleteFoodItem(foodItem: FoodItem): Long
+    suspend fun deleteFoodItem(foodItem: FoodItem): Int
     fun fetchAllFoodItems(): Flow<List<FoodItem>>
     fun fetchAllFoodItemsByName(): Flow<List<FoodItem>>
 }
