+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/interactors/store/DeleteStoreInteractor.kt
@@ -3,5 +3,5 @@ package uz.behzoddev.restaurantpos.domain.interactors.store
 import uz.behzoddev.restaurantpos.data.local.models.FoodStore
 
 interface DeleteStoreInteractor {
-    suspend operator fun invoke(foodStore: FoodStore): Long
+    suspend operator fun invoke(foodStore: FoodStore): Int
 }
