+++ app/src/main/java/uz/behzoddev/restaurantpos/domain/repositories/StoreRepository.kt
@@ -6,7 +6,7 @@ import uz.behzoddev.restaurantpos.data.local.models.FoodStore
 interface StoreRepository {
     suspend fun insertStore(foodStore: FoodStore): Long
     suspend fun updateStore(foodStore: FoodStore): Long
-    suspend fun deleteStore(foodStore: FoodStore): Long
+    suspend fun deleteStore(foodStore: FoodStore): Int
     fun fetchAllStores(): Flow<List<FoodStore>>
     fun fetchAllStoresById(foodStoreId: Int): Flow<FoodStore>
 }
